tests: prplmesh_base_test: add functions for configuring SSIDs
We often need to configure the SSIDs. Add a function that does this,
using the northbound API. For simplicity, only a simple configuration is
supported by this function. Additional configuration can be done by
calling configure_ssid and then nbapi_set to set the additional things.

This should also check that the AP config renew and WSC exchange is done
- to be added later.

This should also check that the BSSes are indeed configured on each
agent - to be added later.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>