agent: Copy the add_profile2_ap_capability tlv adder
Copy the add_profile2_ap_capability TLV adder function implementation to
the capability reporting task since the task needs to add this TLV to
AP Capability Query Report message.
Add this function as public so when the Unified Agent will be completed
the Auto-Configuration task could use the function on the Capability task.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>