tests: certification: add more passing tests
Those were already passing for a long time, but weren't added to the
list of passing tests.

Note that 4.7.6 is a special one. With only one sniffer in the
testbed, 5GL is the only variant we will be able to pass. This is
because the backhaul is also 5GL, so backhaul and fronthaul are on the
same channel and one sniffer is sufficient.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>