bwl: create macro ALLOC_SMART_BUFFER
https://jira.prplfoundation.org/browse/PPM-1020

This is a preparatory commit.

Macro ALLOC_SMART_BUFFER is defined in several files of the BWL library.
A follow-up commit will need it in a file where it has not been
defined yet. Instead of duplicating it yet another time, we better
create it in a common place so it is available to all files in the BWL
and remove it everywhere else. This removal will be done in next commit.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>