agent: backhaul steering: add target bssid as optional argument
We're currently failing MAP-4.9.1 because the target BSSID is not
correct in case it's sent in a response to an invalid request.

Using the BSSID we're currently connected to is only valid in case of
a successful steering. In the other cases, it needs to be provided.

Add an optional 'target_bssid' argument to
'create_backhaul_steering_response()'. It will later be used to
provide the original target BSSID from the backhaul steering request.

If the new argument is not provided, use the BSSID we're currently
connected to, like it was the case before.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>