agent: provide the target BSSID in case of unsuccessful steering
Use the new 'target_bssid' parameter to provide the target bssid when
the steering is unsuccessful.

There was already a 'bssid' variable created further below for
'BACKHAUL_STEERING_REQUEST_REJECTED_TARGET_BSS_SIGNAL_NOT_SUITABLE'. Move
it before checking the channel, so that we can use the same variable
in both cases.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>