agent: backhaul_manager: disconnect on steering is not an error
https://jira.prplfoundation.org/browse/PPM-1020

When a backhaul STA steering request is processed, agent disconnects and
connects back again (to a new BSSID). Currently we are considering a
disconnect event as an error but, when such disconnect event is due to a
steering operation, it is not an error but only a natural side effect.

Modify the handling of the CTRL-EVENT-DISCONNECTED event so it is not
considered an error when it is a consequence of a steering operation.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>