agent: backhaul_manager: send backhaul steering response when connected
https://jira.prplfoundation.org/browse/PPM-1020

Currently, backhaul STA steering response message is sent when agent's
FSM transitions to the CONNECTED state after a backhaul STA steering
request message has been processed.

However, if agent's FSM was in the OPERATIONAL state when the steering
request was received, it will return to the OPERATIONAL state when
steering completes and the backhaul STA steering response message will
not be sent back to controller.

To fix the problem, move the code block that sends the backhaul STA
steering response message from the EState::CONNECTED switch case to the
Event::Connected switch case (that is, send the steering response
whenever the CTRL-EVENT-CONNECTED event is received, no matter in which
state the FSM was).

Signed-off-by: Mario Maz <mmazzapater@gmail.com>