agent: Fix cac types TLV creation
PPM-1055
While adding new element to a list in the TLVF, one must first call
`create_element()` method and after setting all the data call
`add_element()` method.
The code missed the `add_cac_types()` to add it to the TLV.
Add the call to the missing method.

While at it, the call for `add_element()` might fail so check it's
return value.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>