controller: db: add class for BSS
Add a class that represents a BSS.

For now, the  existing represenation of BSSes inside the node is kept in
parallel with the new bss class. Eventually, all vap information should
be removed from the node class.

This is part of refactoring the controller DB to have proper classes for
all node types.

https://jira.prplfoundation.org/browse/PPM-1059

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>