controller: network_map: use new bss class
We need the additional get_bsses method on the radio to support this.

The backhaul_vap flag can't be set with EasyMesh messages, and we don't
really need it, so it won't be part of the bss class. However, the BML
network map still expects it. Therefore, always set it to false.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>