agent: channel scan task: get scan results for request
To get all the scan results for a request a helper function is needed.
Since the request needs to handle multiple radios, and different modes
of operation, a helper functions makes this process simpler.
Thus allowing the send_channel_scan_report_to_controller function to
focus on it's main purpure.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>