agent: channel scan task: refactor send report fragmentation.
To allow for message fragmentation, the send-report function needs to
be refactored.
As part of the refactoring, the report needs to be split up into
smaller manageable sized CMDUs.

We do this by calculating the maximum number of neighbors we can fit
within the current buffer, then create result TLV accordingly.
Once a buffer has been exhausted and cannot fit another TLV with even a
single neighbor (and there are still remaining results to be sent), we
sent the finished CMDU and create another thus freeing the buffer, and
allowing us to continue sending more reports.

When communicating with a prplmesh controller, we send an additional
TLV with a flag that indicates whether the report is done or if there
are more expected reports incoming.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>