controller: handle_dead_node: hoist reported_by_parent out
Preparative commit.

The handle_dead_node function has a parameter hostap_mac. It's unclear
what it was originally intended for, but currently it is used only for
one thing: to determine if the entity that reports the node as dead (=
hostap_mac) is in fact the parent of the node that will be removed.

The problem is that this requires the hostap_mac to still exist in the
database (if it was already removed, it can't be the node's parent).

Since we don't actually need hostap_mac at all, replace it with the
boolean parameter reported_by_parent. In all callers, this parameter is
set as `database.get_node_parent(mac) == hostap_mac`. In most callers,
the hostap_mac argument was anyway already
`database.get_node_parent(mac)` so this is resolved to true.

The call in the topology notification for disassociation looks wrong:
when reported_by_parent is not true, it means that the client has
already associated somewhere else, so the node is not dead, and its
association handling task shouldn't be killed. However, that's not the
subject of this change, so just add a TODO comment.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>