bwl: add scan_type_only
This is a preparatory commit. It will later be used to steer backhaul.

To be able to steer to another backhaul AP, the supplicant has to have
a BSS entry for it. Otherwise, it will fail to steer with:
CTRL_IFACE ROAM: Target AP not found from BSS table

The supplicant create BSS entries whenever it finds an AP during a
scan. Because scanning can be time consuming and can disrupt traffic,
we want to avoid scanning unnecessarily.

Add a new method to scan only on a specific channel, for a specific
BSSID. This will make sure we scan really only what we need before we
try to steer backhaul.

Additionally, use "TYPE=ONLY" to make sure no other action is
performed as the result of the scan. For more information about this
feature, see the commit that adds it in hostapd:
'66fe0f703cb4d360950c1c551dc2691c0788b442'.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>