controller: remove VAPs that are not reported in topology response
The topology response message adds VAPs according to the AP Operational
BSS TLV. However, these VAPs are not removed when they change, e.g. when
a BSS is torn down. This leads to incorrect topology information. It
causes boardfarm tests that use the topology information to fail,
depending on what was configured in earlier tests.

Collect a list of reported_bsses while processing the AP Operational BSS
TLV, and remove any BSSes from the radio if they are not reported. Add a
second variant of remove_vap and add remove_unused_vaps to support this.
Any clients that were connected to the removed VAPs must be cleaned up
as well with handle_dead_node. For this, remove_unused_vaps gets an
additional argument that collects the removed clients. This is pretty
clunky, but it will be improved with further refactoring of the nodes in
the db.

Note that if a radio is not included at all in AP Operational BSS TLV,
its VAPs will be unaffected. The specification is not 100% clear on
this, but it seems that all radios are supposed to be reported in the AP
Operational BSS TLV, so this situation is an error anyway.

https://jira.prplfoundation.org/browse/PPM-1062

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>