agent: backhaul_manager: scan for the target BSSID before steering
Before we attempt to do backhaul steering, we need to make sure the
suppliant has a BSS entry for it. Because it doesn't scan on its own,
do a scan only for the specific bssid and channel.

Since the scan is an asynchronous operation, the steering itself is
moved to the place where we receive the scan results. For this a new
member variable is added to store the channel of the target BSSID. The
operating class is not stored because it's not used after it has been
checked that the channel is valid.

Ideally we should also check when we receive the scan results if the
supplicant found the BSS or not. If not, it means the BSS was not in
range or not on the given channel. In that case we can immediately
reply to the backhaul steering request with a corresponding reason
code. This is however left for later.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>