tests: ClientAssociationDummy: configure SSID before the test
Since the connection map now only contains configured SSIDs, we have to
make sure that the SSID is configured in any test that uses the
connection map. The ClientAssociationDummy test didn't do this, so if a
preceding test left the VAPs unconfigured, the test now fails.

Configure all radios with a single SSID ClientAssocDummy.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>