controller: db: get_node_children: use node bsses
Now we have a list of bsses in the node, we can use that rather than
get_hostap_vaps_bssids to get the node children.

This will allow us to remove hostap_vap_list entirely later.

Note that this will eventually be removed again, when everything is
converted to use the bss class instead of the node hierarchy.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>