framework: accept more formats in mac_from_string
In some cases, we need to read MAC addresses of the form 0xXXXXXXXXXXXX.
Update the mac_from_string functions to accept this format, as well as
XXXXXXXXXXXX. While we're at it, give the function a boolean return
value in case conversion fails, and make sure that in all cases the
address is reset to zero if it fails. Note that most callers still use
the second form which has no error handling other than setting the mac
address to zero.

Add a unit test to validate the existing and the new behaviour.
Unfortunately, the tlvf_test doesn't use googletest yet, so it's
open-coded.

Remove some EOL whitespace in comments (automatic by editor).

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>