prplmesh_utils.sh: start transport after setting ebtables
Recently, an error exit was added if the ebtables command fails.
However, the transport is started just before that, and it gives some
(harmless) ERROR messages before starting up. The result is that the
"Please make sure the ebtables module is loaded" message is a bit
hidden. In addition, the transport keeps running, which makes it look as
if something went wrong with starting the agent.

Fix this by setting up ebtables first. If it fails, the transport is not
started so no messages from it are visible, and the ebtables error is
the last visible error message.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>