tests: boardfarm: remove unnecessary checkpoint
https://jira.prplfoundation.org/browse/PPM-1022

If `configure_ssids()` method is called at the very beginning of a test
then an exception with message "get_packet_capture but no capture file"
is raised. Reason is `configure_ssids()` calls `checkpoint()` and
capture file does not exist yet.

That checkpoint was a bit premature - the idea was to use it to validate
that reconfiguration took place correctly. But since that part isn't
implemented yet, it's better to remove it indeed.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>