controller: refactor handle channel scan report
To align the controller to the newly implemented fragmentation flow,
The current implementation of the Channel-Scan-Report-Message handling
needs to be refactored.

To align with the fragmentation requirements the controller will
validate the received timestamp against our database and see if there
already is an existing record for the given timestamp.
If there isn't we can assume that this is the first fragment of the
report and that the already stored results need to be cleared.
After handling the results we need to verify that the report has ended.
When communicating with a prplmesh agent we sent as additional
"ResultMessageDone" TLV at the end of each report. and if the
"Report Done" within it is set, we can clear the timestamp records and
trigger an event to the task, notifying it that the request was
completed.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>