agent: refactoring abort in-progress ZW-DFS flow unexpected CSA_Finished
The "CSA_Finished" event is raised by the agent's primary radio when a
channel switch is performed into a non-DFS or available DFS channel.

If this event is received unexpectedly, it serves as an indication for
an external channel switch, in which case, the current ZW-DFS flow will
be aborted, releasing the ZW-DFS antenna and give way to the external
channel switch being requested.

If an external channel switch was not detected the ZW-DFS antenna will
still be released, if in use, as part of the CSA_Finished flow.

PPM-1063

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>