agent: retry the ZW-DFS flow on radar hit during CAC
A cac_completed_notification is received when the channel_selection_task
had selected and try to move to a better DFS usable channel.
the move to a DFS channel require a CAC check.

If a RADAR hit is detected during CAC (on the ZW-DFS interface),
cac_completed_notification will be received with success=false.

Current handling: returning the zw-dfs antenna and end the zwdfs flow.

New handling: restart the channel_selection_task flow from
REQUEST_CHANNELS_LIST state - trying to pick another channel which
was not effected by the radar hit.

PPM-1063

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>