agent: release ZWDFS antenna in use when current channel is best
Releasing the ZW-DFS antenna in case where the current failsafe channel
is already second best channel and no farther channel switching is
planed.

This release is needed in case of radar hit during CAC
(on the ZW-DFS interface). If no better channel is found as
"seconde best" the ZW-DFS antenna should not be kept in use.

PPM-1063

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>