ci: move fullimage names to global variables
This is a preparatory commit.

We already define the image names in multiple places. Since they will
also be needed for the release mechanism, move them to global
variables.

There is one place where those new global variables cannot be used
though: in the build jobs. Until [1] is solved we won't be able to
define OPENWRT_IMAGE based on the content of another variable.

[1]: https://gitlab.com/gitlab-org/gitlab-runner/-/issues/1809

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>