agent: platform manager: fix checking if country code is empty
CountryCode is defined as `std::array<char, 2>`.
In f8dd380ee it was checked for being empty with
`std::array::empty()`.
This method checks that the array has non-zero size and
it always returns true.

Country code without any meaningful value assigned to it
contains two '\0' characters and it is what should be
considered empty.

The mistake in checking for being empty results in
dumping content of "empty" country code into logs.
It confuses `grep` utility that marks such files as
binary and treats them differently.
In the end it disrupts testing that is still dependent on
text search in logs.

Fix check for country code emptiness in platform manager.
Fix a typo as well.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>