ucc: listener: pass parameter by const ref
https://jira.prplfoundation.org/browse/PPM-972

This is a preparatory commit.

In a follow-up commit, method `get_send_1905_1_tlv_hex_list` is going to
be used from a method where second parameter `params` is a const ref. To
avoid compilation errors, modify the signature of this method to accept
a const reference instead of a reference.

To allow for this modification, struct `tlv_hex_t` and the methods that
manipulate it have to be modified too. Modification consists in defining
its fields as strings instead of as pointers (to the values included in
`params`).

Signed-off-by: Mario Maz <mmazzapater@gmail.com>