tests: boardfarm: support line number for multi-line log searches
Boardfarm searches log in two steps:
First step is dumping content of selected log files using
`tail` and using boardfarm built-in `expect` function to
look for matches.
Second step is running `tail` with its content piped
through `grep` to add line numbers.

The issue is in the second step. The match from the first
step is put into parentheses and given to `grep`. It does
not work if there are new line characters.

To avoid this, give only the first line of the step 1 match
to `grep`.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>