ucc: listener: define WFA-CA command handlers
https://jira.prplfoundation.org/browse/PPM-972

Currently, the processing of certification commands "dev_reset_default"
and "dev_set_config" is modeled as a state machine and implemented
between the FSM in the backhaul manager and some state variables in
agent's UCC listener. This design has some drawbacks (see PPM-889) and
is very difficult to maintain and evolve. Also, and because of the
dependencies between the two components, code is not easily testable.

To overcome these problems, WFA-CA commands must be entirely processed
in the backhaul manager and the UCC listener's responsibility must be
limited to receive commands and send replies. For this to work, the
backhaul manager needs a mechanism to install callback functions in the
UCC listener to get notified of the reception of such commands and a
function to send asychronous replies.

Define handler functions for "dev_reset_default" and "dev_set_config"
WFA-CA commands received from UCC client during certification tests.
Define functions to register and clear these handlers. Define a function
to send an asynchronous reply to the UCC client once a command has been
processed.

This is a preparatory commit. In a follow-up commit both agent and
controller will implement and install these handlers and the UCC
listener will use them instead of processing the commands itself.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>