builder: rdk-b: use 2021Jan31 tag
Up to now, we were just using the "latest" RDK-B release to build for
Turris Omnia. However, this is changing all the time, which leads to
occasional breakage.

Switch to a tag that is hopefully more stable. This is just a snapshot,
so no "better quality" than anything else - but at least, it shouldn't
be modified any more.

See https://wiki.rdkcentral.com/display/RDK/Jan_31_2021_RDKB_Gateway

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>