boardfarm: pass fail_on_mismatch to _device_wait_for_log
If `fail_on_mismatch` parameter is set to false,
`wait_for_log` method is expected to return values
instead of throwing an exception if given line has not been
found in the logs.

In the code `wait_for_log` parameter was not used.

Pass `fail_on_mismatch` to `_device_wait_for_log`.
Change the logic of the method accordingly.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>