ci: read the UCI radio mode to set channels
The radio number used in the script were specific to a stock Turris
Omnia. This prevented the script from being used on another device
like the Gl.iNet B1300.

Add a new function to set channels based on what the UCI reports for
'hwmode'. A logging line is also added to allow us to make sure of
what was configured on the device.

Also add comments to explain why we set specific channels.

The 'htmode' workaround is also moved to the same function.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>