ci: propagate CI_PIPELINE_SOURCE to the platform child pipeline
We used to use 'if: '$CI_PIPELINE_SOURCE == "schedule"'' as a
rule to build for the NEC device only in scheduled pipelines.

Now that it has been moved to the child pipeline,
'$CI_PIPELINE_SOURCE' will always be "parent_pipeline".

To solve this, set '$CI_PARENT_PIPELINE_SOURCE' in the child pipeline
with the '$CI_PIPELINE_SOURCE' value in the parent pipeline.

Use the new variable for the NEC builds.

Do the same for the '-mmx' builds, which are also supposed to be built
in scheduled pipelines.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>