agent: backhaul_manager: change optional param to required
https://jira.prplfoundation.org/browse/PPM-1083

Parameter `target_bssid` in method `create_backhaul_steering_response`
is optional. However, it is not specified only 1 out of 5 times and the
complexity added to the method to deal with an optional parameter is
not worth the gain obtained.

Change method signature to require target BSSID and simplify
the method as it does not have to deal with an optional parameter
anymore.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>