agent: backhaul_manager: do not err if scan_type_only fails
https://jira.prplfoundation.org/browse/PPM-1083

Do not end with error if call to `scan_type_only` fails.

Add comment describing why it is not an error and why we are not
sending a Backhaul Steering Response message to controller if
call to `scan_type_only` fails.

Also, replace the log severity from ERROR to WARNING when this function
fails.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>