bwl: rename scan_type_only to scan_bss
https://jira.prplfoundation.org/browse/PPM-1083

`scan_type_only` is not a good name, the type=only is hostapd-specific,
and not very understandable. In fact, we would expect all our scans to
be scan only - the `initiate_scan()` function is expected to return
scan results and to have an explicit `connect()` to actually connect to
it.

Therefore, better rename to `scan_bss`.

And while we are at it, improve the method documentation.

Signed-off-by: Mario Maz <mmazzapater@gmail.com>