agent: Separate ethernet struct to WAN and LAN
Until now, prplMesh only took into account the WAN interface as
a wired interface since it was needed for wired backhaul.
Traffic separation feature requires configuring the other LAN ports,
so add them as well to the ethernet struct.

This commit changes the original structure of the WAN port on
the ethernet struct.
The next commit will rename all references to it.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>