bwl: nl80211: bss[0] is an actual VAP
In commit 6dc4c985b7fc4e2fc9c25126cb2a258fc38a929a, the
refresh_vaps_info() function was completely rewritten. As part of this
rewrite, the first BSS was skipped.

This is however not correct for nl80211 - the first VAP is an actual
VAP. It's only on the dwpal drivers that the first VAP is a dummy VAP.

Remove this special case, as well as the accompanying comment.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>