tools: openwrt: bump to latest prplwrt main branch
In prplwrt, the following changes were needed on top of the main
prplwrt branch ('f35e28847dcc8b990ec4525fb04cb84b8392edee'):

- Add AX3000_1600_ETH_11AXUCI in intel_mips profile.

- Bump feed-intel to 'e5757b371ca47e873a0342cfee2f5c89cf095b3d'.

- Add the patch for 4addr mode from
https://gitlab.com/prpl-foundation/prplwrt/prplwrt/-/merge_requests/64.

- Bump feed-prpl to 'ee7c1bbce5209cc7b6f12c01a99759c45528f506' to
bring in the latest changes we did there (and that we already used in
this repository).

- Add missing debugging packages to the debug profile.

In this repository, the changes described below were needed.

With recent versions of prplWrt, the hashes of the feed-prpl and
feed-intel are included in the profile files. This means we no longer
have to keep track of them in this repository.

feed_intel is now in 'profiles/intel_mips.yml'.

feed_prpl is in 'profiles/prpl.yml'.

Remove the unnecessary feeds from this repository. Because the
netgear-rax40 and axepoint profiles do not exist anymore in prplWrt,
use intel_mips instead. With this change we can merge all the CI jobs
that build for Intel platforms into one (this will be done in a
follow-up commit).

Since we now use the prpl profile directly, append the prplMesh
package to it. Eventually prplMesh should be contained in the prpl
profile by default, but it's not the case yet. Since it has two
variants and the profile has to use the correct one by default, we
first need to test how we can handle it. This will be done in
PPM-1112.

The debug profile contains 'iw-full', but intel_mips depends on
iwlwav-iw which clashes with the original 'iw-full'. For intel_mips,
just remove it from the profile.

Finally, for the nl80211 builds 'autoconf-archive' is now needed, or
autoconf will fail with 'error: possibly undefined macro: AX_PTHREAD'.

All of this will allow us to more easily switch to using profile files
for the non-Intel platforms as well (in a follow-up commit).

Fixes PPM-1080

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>