tools: openwrt: use profile files to build for non-Intel platforms
Move the configuration for the Intel platforms out of the case
statement. The same code is now used to configure the nl80211-based
platforms, so remove the case statement entirely.

The parts that were specific to building the Turris Omnia and
Gl.iNet B1300 (i.e. building without a profile file) have also
been removed.

Since PRPLMESH_VARIANT was missing for the Intel platforms and because
we use it for the gen_config.py profile, add it.

There is one thing that is platform-specific and we have to keep: the
removal of iw for intel_mips (since intel_mips depends on iwlwav-iw).

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>