controller: db: add sAgent struct
The sAgent struct and its inner structures will replace the node class.

The Station class will be added later, but a reference to it is already
included in this definition.

sAgent currently only contains the basic infrastructure, additional
fields will be migrated from the node class in later commits.

For the transition, we maintain a parallel hierarchy with the node
hierarchy. When adding agents and radios, they are added both as sAgent
and in the node hierarchy. BSSes will be handled in later commits, they
are a bit more complicated.

Note that the controller code currently never removes agent or radio
nodes.

https://jira.prplfoundation.org/browse/PPM-1057
-
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>