bpl: support prplWrt and UGW libbridge in same platform type
Recent changes that introduced the UGW 8.4.2 support to prplWrt changed
the BPL cmake and bpl_arp.cpp in a way that breaks the UGW and RDKB
compilation.

The change is added to support linking the libbridge directly from the
install-dir, but currently the WAVE UGW and RDKB do not install it and
it is still required to look it up in the build-dir.

Added support for both by prefering the install-dir and if not exist use
the build-dir.
Due to library naming differences and until a permanent fix is
introduced, a new LIB_BRIDGE_UGW flag is introduced to determine in
bpl_arp the library name to include.

Signed-off-by: Adam Dov <adamx.dov@intel.com>
(cherry picked from commit d2d74fecb1fc15458e344ea3137c2919528b8b88)
Signed-off-by: Frederik Van Bogaert <frederik.vanbogaert@mind.be>