ci: merge builds for Intel devices
The intel_mips profile is now building images for all the Intel
devices.

Merge all the build jobs for Intel devices into
"build-for-intel-mips".

All the jobs that depended on a build for a specific Intel device now
depend on "build-for-intel-mips".

Our scripting depends on the prplMesh ipk and fullimages to be under
'build/device_name'. With these changes, for the Intel platforms the
builds will now instead be under build/intel_mips. Since this will be
much easier to resolve once PPM-1121 is done, for now just create
symlinks for the rax40 and axepoint with a TODO. The artifacts path is
also changed to include everything under 'build/'.

Note that the NEC builds were previously only done in scheduled
pipelines, with this approach they will be done along with the other
ones.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>