tools: openwrt: build a fixed revision as a base
Up to now whenever we made even a small change in prplWrt, it lead to
a complete rebuilt of the prplWrt tree, which resulted in very long
build times.

To try to reduce the build time, first checkout and build a fixed
revision of prplWrt. The point is that this base revision would only
be updated when required (when OPENWRT_VERSION switches to a new major
release for example). As a consequence, for most builds where
OPENWRT_VERSION changes we would keep a built toolchain and most
packages for the target, which should reduce the build times.

This also means the size of each docker image will increase even
more. However for the build servers, the overall disk usage should
decrease due to the fact that Docker re-uses layers between
images. With these changes only the last layers (after the base tree
is built) - which are smaller - will differ.

TARGET_SYSTEM is moved before the base prplWrt build, because we need
it as an argument to gen_config.py.

Fixes PPM-1122

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>