ci: run MAP-4.2.2_BH24G in the merge train
Up to now we didn't have any onboarding test with the nl80211 variant.

Run the wireless onboarding test 4.2.2, for each merge train pipeline.

Because we have a number of stability issues which affect WPS (see for
example PPM-1104, PPM-96, PPM-956, PPM-1105) do as we do for the
axepoint: retry the job up to 2 times.

Since in the meantime on Gitlab side [1] is fixed, this will really
take effect in the merge train.

[1]: https://gitlab.com/gitlab-org/gitlab/-/issues/244856

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
(cherry picked from commit a326ab02921bc4df3fa240102112453ffe09371f)
Signed-off-by: Frederik Van Bogaert <frederik.vanbogaert@mind.be>