controller: remove VAPs that are not reported in topology response
The topology response message adds VAPs according to the AP Operational
BSS TLV. However, these VAPs are not removed when they change, e.g. when
a BSS is torn down. This leads to incorrect topology information. It
causes boardfarm tests that use the topology information to fail,
depending on what was configured in earlier tests.

Use mac_map's keep_new functionality to get a list of removed BSSes
while adding the new ones.

Note that if a radio is not included at all in AP Operational BSS TLV,
its VAPs will be unaffected. The specification is not 100% clear on
this, but it seems that all radios are supposed to be reported in the AP
Operational BSS TLV, so this situation is an error anyway.

https://jira.prplfoundation.org/browse/PPM-1062

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>