beerocks fsm: return a reference to local variable
The handle() function has two exit (return) points.
In one of the a reference to a member of a local variable
is returned. This commit changes the return type to be by value
instead of by a reference, and thus solves the problem.

The caveat of this is that if at some point the state type
(TState) would be used with a large object, and not as an enum,
we may pay the copy penalty (unless RVO is done by the compiler)

Signed-off-by: Ran Regev <ran.regev@devalore.com>
(cherry picked from commit 6ab1642998be51022249ad7f34219dda4487b7e6)
Signed-off-by: Frederik Van Bogaert <frederik.vanbogaert@mind.be>