tests: boardfarm: ensure rax40 serial device does not change
One of the configuration parameters of boardfarm is
`conn_cmd`. It is a command to connect used to the device
console.

It was set to `cu -s 115200 -l /dev/ttyUSB0 -f`.
The build server has several serial devices, so the number
after `ttyUSB` can (and did) change after reboot.

Use `/dev/netgear-rax40-1` as the serial device name.
It is configured to always point to the RAX40 router.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>