controller:topology_task: send topology query
The Topology Query message is triggered only in two situations:
when an autoconfiguration request is received, and when a topology notification is received.
This is not sufficient. It additionally must send a topology query when it receives
a topology response, if it contains any neighbor we didn't know about yet.

Send a Topology Query when new neigbor was reported.

PPM-927
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>