tests: uci: check if a wifi configuration is in progress
When the device boots for the first time after an upgrade, our script
regenerating the wifi configuration can in some cases conflict with
the modifications the device itself does.

It can for example be seen in [1] where the configuration ended up
with a missing 'path' for radio0.

In [2], it's the "type" parameter for radio1 that was missing.

If 'wifi' is running on the device, just wait some more until we
re-generate the wireless configuration.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1036999676
[2]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/1036999822

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>