son_slave: clear the channels list on slave init
On country configuration, the son-slave is being reset but the prplmesh
itself isn't - which causes the DB to still contatin the old country
channel-list. This is since (for optimization) the map is not cleared on
each call to unify_channels_list().

It was seen that due to residue from old country channels-list, a
switch-channel for invalid channel was performed by the ZWDFS flow.

Clear the channels-list on slave init.

Signed-off-by: Adam Dov <adamx.dov@intel.com>