tools: openwrt: move amx to a SAH profile
packages-rax40.yml contains the packages that are needed to build
Ambiorix, but nothing specific to the rax40.

The feed that is needed for amx however is contained in
netgear-rax40.yml, while the rax40 might not be the only platform for
which we want to build Ambiorix.

Like it's done for mmx, move amx to its own profile that contains the
feed and packages. Since the feed is called 'sah', use the same name
for the file.

In a follow-up commit, it will also be used to build Ambiorix for
non-intel platforms.

While at it, add the profiles we use to 'prplwrt-version'. This is
required to make sure our deploy_firmware script knows when the
firmware needs to be updated.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>