agent: Fix getting to STOPPED state on GW after slave reset
On GW Platform, we clear the WAN interface from the database, once getting the
configuration from the Platform Manager. Since we initialize the local_gw flag later,
add a check if the WAN interface is empty instead of checking the local_gw flag.

This change solves a scenario that causes the son_slave to move to the STOPPED state,
that happens because on a later state the WAN interface name is cleared from the
database if the local_gw flag is set which triggers failure on "linux_iface_get_mac()" and
increment the stop_on_failure counter.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>