agent: Comment out code snippet
There is a code snippet on the OPERATIONAL state of the Backhaul Manager
that has been commented out. This snippet is responsible to sample the WAN
Monitor state periodically.
There is another part of that code on the CONNECTED state as well, that was
not commented out, although it is not active.
Comment it as well.

Since the code snippets are using the WAN interface, do the same wrapping
as has been done on the previous commit, so when this code will become
operational again, it will be correct.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>